remove cli code, move to new repo
