+++ api/oapi-models-codegen.yaml
@@ -1,0 +1,4 @@
+package: models
+output: models/models.gen.go
+generate:
+  models: true
