+++ api/oapi-server-codegen.yaml
@@ -1,0 +1,8 @@
+package: api
+output: server.gen.go
+additional-imports:
+  - package: raven/api/models
+    alias: .
+generate:
+  echo-server: true
+  embedded-spec: true
