+++ makefile
@@ -1,0 +1,9 @@
+build:
+	go generate ./...
+	go build -o main server.go
+
+run: build
+	./main
+
+watch:
+	air
