+++ openapi.yaml
@@ -1,0 +1,15 @@
+openapi: 3.1.0
+info:
+  version: 0.0.1
+  title: Raven API
+
+paths:
+  /v1/api/events:
+    post:
+      tags: -openapi
+      operationId: ingestEvents
+      requestBody:
+        content:
+          application/json:
+            schema:
+              type: object
